+++ MelonLoader/Melons/Mod.cs
@@ -5,14 +5,24 @@ namespace MelonLoader
     public abstract class MelonMod : MelonBase
     {
         /// <summary>
-        /// Runs when a Scene has Loaded.
+        /// Runs when a Scene has Loaded and is passed the Scene's Build Index.
         /// </summary>
         public virtual void OnSceneWasLoaded(int buildIndex) { }
 
         /// <summary>
-        /// Runs when a Scene has Initialized.
+        /// Runs when a Scene has Loaded and is passed the Scene's Name.
+        /// </summary>
+        public virtual void OnSceneWasLoaded(string sceneName) { }
+
+        /// <summary>
+        /// Runs when a Scene has Initialized and is passed the Scene's Build Index.
         /// </summary>
         public virtual void OnSceneWasInitialized(int buildIndex) { }
+
+        /// <summary>
+        /// Runs when a Scene has Initialized and is passed the Scene's Name.
+        /// </summary>
+        public virtual void OnSceneWasInitialized(string sceneName) { }
 
         /// <summary>
         /// Can run multiple times per frame. Mostly used for Physics.